% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_StringsBasics.R
\name{+}
\alias{+}
\title{String Concatenation}
\usage{
x + ...
}
\arguments{
\item{x}{A character vector.}

\item{...}{Additional strings to concatenate.}
}
\description{
A wrapper around \code{paste0()} for concatenating strings.
}
